Still attaching all the functionality to create a runnable system. (code: X) Fixing AffinityFunction..

+++ src/main/java/edu/colostate/cs/fa2017/stretch/util/FileEditor.java
@@ -8,38 +8,62 @@ public class FileEditor {
     String fileName;
     String placeHolder;
     String replacement;
+    String group;
 
-    public FileEditor(String fileName, String placeHolder, String replacement){
+    public FileEditor(String fileName, String placeHolder, String replacement, String group){
 
         this.fileName = fileName;
         this.placeHolder = placeHolder;
         this.replacement = replacement;
+        this.group = group;
     }
 
-    private void replace() {
 
-        String oldFileName = fileName;  //"./config/util/test.xml";
-        String tmpFileName = fileName+"tmp"; //"./config/util/test1.xml";
+    public String replace() {
+
+        String oldFileName = fileName;  //"./config/util/template.xml";
+        String tmpFileName = fileName.replaceAll("Template", group); //"./config/util/test1.xml";
+
+
+        //Delete the config file, if exist
+        File oldFile = new File(tmpFileName);
+        if(oldFile.isFile()){
+            oldFile.delete();
+        }
+
 
         BufferedReader br = null;
         BufferedWriter bw = null;
         try {
+            //Assuming the order is as in the template..groupname first, then donated,...
+            String[] place = placeHolder.split("$$");
+            String[] replace = replacement.split("##");
+
+            if(place.length!=replace.length){
+                return "ERROR";
+            }
+
+            int len = place.length;
+            int index = 0;
+
             br = new BufferedReader(new FileReader(oldFileName));
             bw = new BufferedWriter(new FileWriter(tmpFileName));
             String line;
             while ((line = br.readLine()) != null) {
-                if (line.contains(placeHolder))
-                    line = line.replace(placeHolder, replacement);
+                if (index < len && line.contains(place[index])){
+                    line = line.replace(place[index], replace[index]);
+                    index++;
+                }
                 bw.write(line+"\n");
             }
         } catch (Exception e) {
-            return;
+            System.out.println(e);
         } finally {
             try {
                 if(br != null)
                     br.close();
             } catch (IOException e) {
-                //
+                System.out.println(e);
             }
             try {
                 if(bw != null)
@@ -49,12 +73,14 @@ public class FileEditor {
             }
         }
         // Once everything is complete, delete old file..
-        File oldFile = new File(oldFileName);
-        oldFile.delete();
+//        File oldFile = new File(oldFileName);
+//        oldFile.delete();
 
         // And rename tmp file's name to old file name
-        File newFile = new File(tmpFileName);
-        newFile.renameTo(oldFile);
+//        File newFile = new File(tmpFileName);
+//        newFile.renameTo(oldFile);
+
+        return tmpFileName;
 
     }
 }
